+++ src/main/java/com/kustacks/kuring/admin/presentation/AdminControllerV1.java
@@ -15,9 +15,9 @@ import com.kustacks.kuring.common.dto.ResponseDto;
 import com.kustacks.kuring.common.exception.APIException;
 import com.kustacks.kuring.common.exception.code.ErrorCode;
 import com.kustacks.kuring.common.exception.InternalLogicException;
-import com.kustacks.kuring.common.firebase.FirebaseService;
-import com.kustacks.kuring.common.firebase.exception.FirebaseInvalidTokenException;
-import com.kustacks.kuring.common.firebase.exception.FirebaseMessageSendException;
+import com.kustacks.kuring.message.firebase.FirebaseService;
+import com.kustacks.kuring.message.firebase.exception.FirebaseInvalidTokenException;
+import com.kustacks.kuring.message.firebase.exception.FirebaseMessageSendException;
 import com.kustacks.kuring.feedback.domain.Feedback;
 import com.kustacks.kuring.notice.domain.Notice;
 import com.kustacks.kuring.user.domain.User;
