+++ src/main/java/com/kustacks/kuring/feedback/presentation/FeedbackControllerV1.java
@@ -2,7 +2,7 @@ package com.kustacks.kuring.feedback.presentation;
 
 import com.kustacks.kuring.common.exception.APIException;
 import com.kustacks.kuring.common.exception.code.ErrorCode;
-import com.kustacks.kuring.common.firebase.FirebaseService;
+import com.kustacks.kuring.message.firebase.FirebaseService;
 import com.kustacks.kuring.feedback.business.FeedbackService;
 import com.kustacks.kuring.feedback.common.dto.SaveFeedbackV1Request;
 import com.kustacks.kuring.feedback.common.dto.SaveFeedbackResponse;
