+++ src/main/java/com/kustacks/kuring/message/firebase/exception/FirebaseInvalidTokenException.java
@@ -1,4 +1,4 @@
-package com.kustacks.kuring.common.firebase.exception;
+package com.kustacks.kuring.message.firebase.exception;
 
 import com.kustacks.kuring.common.exception.BusinessException;
 import com.kustacks.kuring.common.exception.code.ErrorCode;
