+++ src/main/java/com/kustacks/kuring/message/firebase/exception/FirebaseSubscribeException.java
@@ -1,4 +1,4 @@
-package com.kustacks.kuring.common.firebase.exception;
+package com.kustacks.kuring.message.firebase.exception;
 
 import com.kustacks.kuring.common.exception.BusinessException;
 import com.kustacks.kuring.common.exception.code.ErrorCode;
