+++ src/main/java/com/kustacks/kuring/user/facade/UserCommandFacade.java
@@ -5,9 +5,9 @@ import com.kustacks.kuring.worker.event.SubscribedRollbackEvent;
 import com.kustacks.kuring.notice.domain.CategoryName;
 import com.kustacks.kuring.common.exception.APIException;
 import com.kustacks.kuring.common.exception.code.ErrorCode;
-import com.kustacks.kuring.common.firebase.FirebaseService;
-import com.kustacks.kuring.common.firebase.exception.FirebaseSubscribeException;
-import com.kustacks.kuring.common.firebase.exception.FirebaseUnSubscribeException;
+import com.kustacks.kuring.message.firebase.FirebaseService;
+import com.kustacks.kuring.message.firebase.exception.FirebaseSubscribeException;
+import com.kustacks.kuring.message.firebase.exception.FirebaseUnSubscribeException;
 import com.kustacks.kuring.feedback.business.FeedbackService;
 import com.kustacks.kuring.notice.domain.DepartmentName;
 import com.kustacks.kuring.user.business.UserService;
